Fix SeqPage spamming set_trigleds due to uint64_t/uint16_t comparison
led_mask and locks_on_step_mask_ are uint64_t but trigled_mask and
locks_on_step_mask are uint16_t. After shed_mask on tracks longer than
16 steps, bits 16+ remain set in led_mask, so the change-guard
comparison was always true, causing MD.set_trigleds to fire every
draw frame.

+++ src/mcl/MCL/SeqPage.cpp
@@ -520,7 +520,7 @@ void SeqPage::draw_mask(uint8_t offset, uint8_t device,
     active_track.get_mask(&locks_on_step_mask_, MASK_LOCKS_ON_STEP);
     shed_mask(locks_on_step_mask_, active_track.length, offset);
 
-    if (led_mask != trigled_mask) {
+    if ((uint16_t)led_mask != trigled_mask) {
       trigled_mask = led_mask;
       MD.set_trigleds(trigled_mask, TRIGLED_STEPEDIT);
       if (mask_type == MASK_MUTE) {
@@ -528,7 +528,7 @@ void SeqPage::draw_mask(uint8_t offset, uint8_t device,
         GUI_hardware.led.set_trigleds(mask, TRIGLED_STEPEDIT, 1);
       }
     }
-    if (locks_on_step_mask_ != locks_on_step_mask) {
+    if ((uint16_t)locks_on_step_mask_ != locks_on_step_mask) {
       locks_on_step_mask = locks_on_step_mask_;
       MD.set_trigleds(locks_on_step_mask, TRIGLED_STEPEDIT, 1);
     }
